Move file and fix exams-list.json not pre-existing
