+++ server/api/controllers/exam-bank-controller.ts
@@ -2,6 +2,7 @@ import { Dirent, readdirSync } from "fs";
 import { Exam } from "../../types/exam-bank";
 import { Logger } from "../../util/logger";
 import { ReadWriteController } from "./read-write-controller";
+import fs from "fs";
 
 export class ExamBankController {
   static logger = new Logger("Exam Bank Controller");
@@ -12,6 +13,12 @@ export class ExamBankController {
    */
   static rewriteFile(): void {
     const url = "_hidden/exams-list";
+    const fullPath = `server/data/${url}.json`;
+
+    if (!fs.existsSync(fullPath)) {
+      fs.writeFileSync(fullPath, "");
+    }
+
     ReadWriteController.overwriteJSONDataPath(
       url,
       (statusCode: number) => {
