+++ server/util/directory-prebuilder.ts
